Rename instructions to align with DWKit forms
