+++ source/SIL.AppBuilder.Portal/src/routes/(authenticated)/tasks/[product_id]/instructions/App_Configuration.svelte
