+++ source/SIL.AppBuilder.Portal/src/routes/(authenticated)/tasks/[product_id]/instructions/Authors_Upload.svelte
