+++ source/SIL.AppBuilder.Portal/src/routes/(authenticated)/tasks/[product_id]/instructions/GooglePlay_Configuration.svelte
