+++ source/SIL.AppBuilder.Portal/src/routes/(authenticated)/tasks/[product_id]/instructions/Synchronize_Data.svelte
