+++ source/SIL.AppBuilder.Portal/src/routes/(authenticated)/tasks/[product_id]/instructions/index.ts
@@ -1,31 +1,33 @@
-import APVerify from "./APVerify.svelte";
-import Configuration from "./Configuration.svelte";
-import CreateEntry from "./CreateEntry.svelte";
-import Download from "./Download.svelte";
-import GPConfiguration from "./GPConfiguration.svelte";
-import GPVerify from "./GPVerify.svelte";
-import MakeItLive from "./MakeItLive.svelte";
-import Pending from "./Pending.svelte";
-import ReadinessCheck from "./ReadinessCheck.svelte";
-import SynchronizeData from "./SynchronizeData.svelte";
-import Upload from "./Upload.svelte";
-import Verify from "./Verify.svelte";
+import Asset_Package_Verify_And_Publish from "./Asset_Package_Verify_And_Publish.svelte";
+import App_Configuration from "./App_Configuration.svelte";
+import Create_App_Entry from "./Create_App_Entry.svelte";
+import Authors_Download from "./Authors_Download.svelte";
+import GooglePlay_Configuration from "./GooglePlay_Configuration.svelte";
+import GooglePlay_Verify_And_Publish from "./GooglePlay_Verify_And_Publish.svelte";
+import Make_It_Live from "./Make_It_Live.svelte";
+import Approval_Pending from "./Approval_Pending.svelte";
+import Readiness_Check from "./Readiness_Check.svelte";
+import Synchronize_Data from "./Synchronize_Data.svelte";
+import Authors_Upload from "./Authors_Upload.svelte";
+import Verify_And_Publish from "./Verify_And_Publish.svelte";
 import Waiting from "./Waiting.svelte";
-import WVerify from "./WVerify.svelte";
+import Web_Verify from "./Web_Verify.svelte";
 
-export const instructions: {[key: string]: typeof Configuration} = {
-  "ap_verify": APVerify,
-  "configuration": Configuration,
-  "create_entry": CreateEntry,
-  "download": Download,
-  "gp_configuration": GPConfiguration,
-  "gp_verify": GPVerify,
-  "make_it_live": MakeItLive,
-  "pending": Pending,
-  "readiness_check": ReadinessCheck,
-  "synchronize_data": SynchronizeData,
-  "upload": Upload,
-  "verify": Verify,
+export const instructions: {[key: string]: typeof Waiting} = {
+  "asset_package_verify_and_publish": Asset_Package_Verify_And_Publish,
+  "app_configuration": App_Configuration,
+  // may need to add "author_app_configuration" later
+  // in S1 it has the same instructions as "app_configuration"
+  "create_app_entry": Create_App_Entry,
+  "authors_download": Authors_Download,
+  "googleplay_configuration": GooglePlay_Configuration,
+  "googleplay_verify_and_publish": GooglePlay_Verify_And_Publish,
+  "make_it_live": Make_It_Live,
+  "approval_pending": Approval_Pending,
+  "readiness_check": Readiness_Check,
+  "synchronize_data": Synchronize_Data,
+  "authors_upload": Authors_Upload,
+  "verify_and_publish": Verify_And_Publish,
   "waiting": Waiting,
-  "w_verify": WVerify
+  "web_verify": Web_Verify
 };
